requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
beautifulsoup4==4.11.1
    # via django-bootstrap-v5
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
cryptography==39.0.0
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.1.5
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-bootstrap-v5
    #   djangorestframework
django-allauth==0.52.0
    # via -r requirements.in
django-bootstrap-v5==1.0.11
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
idna==3.4
    # via requests
oauthlib==3.2.2
    # via requests-oauthlib
pillow==9.4.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
pytz==2022.7
    # via djangorestframework
requests==2.28.1
    # via
    #   -r requirements.in
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
urllib3==1.26.13
    # via requests
